Update ConnectFour.java
Added a calculateDraw() method that can be called after the calculateWin method.  If calculateWin ==false, call the draw method.  Will return true or false.  We could also keep track of the number of pieces played and if it hits 42 without a win, then it's a draw.  I'm not big on relying on counters though since this is only a 6x7 array it's literally milliseconds to check the board for a draw.  We don't have to use it but there if we need it.